requirements.txt
Django==3.2.17
djangorestframework==3.14.0
drf-yasg==1.21.5
python-dotenv==0.21.1
